+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 -   Improvements
     -   Changed `{context}.index` to `{context}.sortOrder`.
     -   Added another variable to the `onClick()` and `onCombine()` action tags to return a context.
+    -   Added `onAnyPlayerContextEnter` to trigger on every bot when a player joins a context and changed `onPlayerContextEnter` to trigger on the player bot that joins a context.
 
 ## V0.9.21